Move methods to UserDataRepository that handling lastProcessTekTimestamp.
